+++ daemon/src/main/kotlin/dev/krud/boost/daemon/configuration/application/ro/ApplicationRO.kt
@@ -13,7 +13,7 @@ class ApplicationRO(
     val color: String = DEFAULT_COLOR,
     var effectiveColor: String = DEFAULT_COLOR,
     val icon: String? = null,
-    val sort: Int? = null,
+    val sort: Double? = null,
     val parentFolderId: UUID? = null,
     var health: ApplicationHealthRO = ApplicationHealthRO.unknown()
 )
